[API] Add empty files (controller/service/repository) for new models

+++ src/main/java/server/controller/ArticleController.java
@@ -5,11 +5,11 @@ import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import server.model.Article;
 import server.service.ArticleService;
+import server.service.ClientService;
+
 import java.util.List;
 import static org.springframework.http.HttpStatus.OK;
-import static org.springframework.web.bind.annotation.RequestMethod.DELETE;
-import static org.springframework.web.bind.annotation.RequestMethod.GET;
-import static org.springframework.web.bind.annotation.RequestMethod.POST;
+import static org.springframework.web.bind.annotation.RequestMethod.*;
 
 
 @RestController
@@ -19,14 +19,17 @@ public class ArticleController {
     @Autowired
     private ArticleService articleService;
 
+    @Autowired
+    private ClientService clientService;
+
     @RequestMapping( path = "/all", method = GET)
     @ResponseStatus(value = OK)
     public List<Article> getAllArticles(){
-        List<Article> allArticles = articleService.getAllArticles();
-        return allArticles;
+        List<Article> listArticles = articleService.getAllArticles();
+        return listArticles;
     }
 
-    @RequestMapping( value = "/{id}", method = GET)
+    @RequestMapping(method = GET)
     @ResponseStatus(value = OK)
     public Article getArticleById(@PathVariable int id){
         return articleService.getArticleById(id);
@@ -34,16 +37,27 @@ public class ArticleController {
 
     @RequestMapping(method = POST)
     @ResponseStatus(HttpStatus.OK)
-    public void addArticle(@RequestBody Article article){
-        articleService.addArticle(article);
+    public void addArticle(@RequestBody Article article, @RequestParam("token") String token){
+        if (clientService.adminAccess(token)){
+            articleService.addArticle(article);
+        }
     }
 
-    @RequestMapping( value = "/delete/{id}", method = DELETE)
+    @RequestMapping(method = PUT)
+    @ResponseStatus(HttpStatus.OK)
+    public void updateArticle(@RequestBody Article article, @RequestParam("token") String token){
+        if (clientService.adminAccess(token)){
+            articleService.updateArticle(article);
+        }
+    }
+
+    @RequestMapping(method = DELETE)
     @ResponseStatus(HttpStatus.OK)
     @ResponseBody
-    public void deleteArticle(int id){
-        articleService.deleteArticle(id);
+    public void deleteArticle(@RequestParam("id") int id, @RequestParam("token") String token){
+        if (clientService.adminAccess(token)){
+            articleService.deleteArticle(id);
+        }
     }
-
 
 }
